+++ i18n/es/readme/CHANGELOG.md
@@ -2,6 +2,7 @@
 
     + Added shortcut for pushing reference to TeXstudio with Better BibTeX
     + Add "Advanced search" shortcut
+    + Add "Locate with Google Scholar" shortcut
 
 * In version 3.3.3: